+++ crates/astria-auctioneer/CHANGELOG.md
@@ -12,3 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial release [#1839](https://github.com/astriaorg/astria/pull/1839).
+
+### Changed
+
+- Use v2 execution API instead of v1 [#2006](https://github.com/astriaorg/astria/pull/2006).